Windows port: last ttnn MSVC fixes -- ccl fusion header, fabric socket gates
- ccl_host_datastructures.hpp: explicit narrowing casts in compile-time args
- fabric_socket / bidirectional_fabric_socket: send/recv route through the
  experimental ccl send_async/recv_async ops, which are outside the trimmed
  subset; under TTNN_TRIMMED_BUILD they throw instead (single-host M0 scope)
- where_tss visit lambda: cast the mixed-type scalar pair explicitly (MSVC
  hard-errors on narrowing inside the braced vector init)

(cherry picked from commit f23ae7dc7d65bf365b14bfa19881366df7b4171a)

+++ ttnn/cpp/ttnn/operations/ccl/ccl_host_datastructures.hpp
@@ -258,8 +258,8 @@ public:
             this->termination_mode,
             1,
             static_cast<uint32_t>(this->num_senders > 0 && active_channels.at(0).is_sender),
-            this->num_buffers_per_channel,
-            chip_id};
+            static_cast<uint32_t>(this->num_buffers_per_channel),
+            static_cast<uint32_t>(chip_id)};
     }
 
     [[nodiscard]]
